requirements.txt
pyyaml==2.4.4
requests==2.22.0